# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LabDifficulty = typeof LabDifficulty[keyof typeof LabDifficulty];


export const LabDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Lab {
  id: string;
  title: string;
  /** Technology track (e.g. "linux", "terraform") */
  track: string;
  /** Skill tier within the track (1 = foundation, 2 = intermediate, 3 = advanced) */
  level: number;
  category: string;
  difficulty: LabDifficulty;
  summary: string;
  estimatedMinutes: number;
  order: number;
}

export interface LabTask {
  id: string;
  description: string;
}

export type LabDetail = Lab & {
  /** Markdown instructions for the lab */
  instructions: string;
  objectives: string[];
  tasks: LabTask[];
  /** Terminal targets this lab exposes (e.g. ["server1","server2"]) */
  terminals: string[];
};

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  none: 'none',
  starting: 'starting',
  running: 'running',
  stopped: 'stopped',
  error: 'error',
} as const;

export interface Session {
  labId: string;
  status: SessionStatus;
  terminals: string[];
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export interface CheckResult {
  id: string;
  label: string;
  passed: boolean;
  message: string;
}

export interface VerifyResult {
  labId: string;
  passed: boolean;
  score: number;
  checks: CheckResult[];
  verifiedAt: string;
}

export type ProgressEntryStatus = typeof ProgressEntryStatus[keyof typeof ProgressEntryStatus];


export const ProgressEntryStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  passed: 'passed',
} as const;

export interface ProgressEntry {
  labId: string;
  status: ProgressEntryStatus;
  bestScore: number;
  /** @nullable */
  lastAttemptAt?: string | null;
}