# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /labs/{labId}:
    get:
      operationId: getLab
      tags: [labs]
      summary: Get lab detail
      parameters:
        - name: labId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lab detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    Lab:
      type: object
      required:
        [id, title, track, level, category, difficulty, summary, estimatedMinutes, order]
      properties:
        id:
          type: string
        title:
          type: string
        track:
          type: string
          description: Technology track (e.g. "linux", "terraform")
        level:
          type: integer
          description: Skill tier within the track (1 = foundation, 2 = intermediate, 3 = advanced)
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        summary:
          type: string
        estimatedMinutes:
          type: integer
        order:
          type: integer
    LabTask:
      type: object
      required: [id, description]
      properties:
        id:
          type: string
        description:
          type: string
    LabDetail:
      allOf:
        - $ref: "#/components/schemas/Lab"
        - type: object
          required: [instructions, objectives, tasks, terminals]
          properties:
            instructions:
              type: string
              description: Markdown instructions for the lab
            objectives:
              type: array
              items:
                type: string
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/LabTask"
            terminals:
              type: array
              description: Terminal targets this lab exposes (e.g. ["server1","server2"])
              items:
                type: string
...