# lib/api-zod/src/generated/types/lab.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LabDifficulty } from './labDifficulty';

export interface Lab {
  id: string;
  title: string;
  /** Technology track (e.g. "linux", "terraform") */
  track: string;
  /** Skill tier within the track (1 = foundation, 2 = intermediate, 3 = advanced) */
  level: number;
  category: string;
  difficulty: LabDifficulty;
  summary: string;
  estimatedMinutes: number;
  order: number;
}